perf(evm): keep the DAO transition out of the zkEVM guest
ApplyDaoTransition names the Dao fork, and naming any fork compiles its
whole ancestor chain into the image. The guest serves post-merge blocks,
so the transition can never apply there.

The method splits along the existing BlockProcessor.std.cs and
BlockProcessor.zkevm.cs partials. The standard build is unchanged. The
guest keeps the block-number check and throws on a match, so a block
needing the transition is refused rather than processed without it.

Guest .text falls 4,296,656 -> 4,281,712 bytes and ROM from 32.35% to
32.24% of capacity, for the same state root and the same step count.
