+++ src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.std.cs
@@ -1,8 +1,12 @@
-// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
+﻿// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System.Runtime.CompilerServices;
+using Nethermind.Blockchain;
 using Nethermind.Core;
+using Nethermind.Evm.State;
+using Nethermind.Int256;
+using Nethermind.Specs.Forks;
 
 namespace Nethermind.Consensus.Processing;
 
@@ -14,4 +18,32 @@ public partial class BlockProcessor
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private static partial bool ShouldCalculateReceiptsInBackground(TxReceipt[] receipts) =>
         receipts.Length >= BackgroundReceiptCountThreshold || CountLogs(receipts) >= BackgroundLogCountThreshold;
+
+    /// <inheritdoc/>
+    private partial void ApplyDaoTransition(Block block)
+    {
+        ulong? daoBlockNumber = _specProvider.DaoBlockNumber;
+        if (daoBlockNumber.HasValue && daoBlockNumber.Value == block.Header.Number)
+        {
+            ApplyTransition();
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        void ApplyTransition()
+        {
+            if (_logger.IsInfo) _logger.Info("Applying the DAO transition");
+            Address withdrawAccount = DaoData.DaoWithdrawalAccount;
+            if (!_stateProvider.AccountExists(withdrawAccount))
+            {
+                _stateProvider.CreateAccount(withdrawAccount, 0);
+            }
+
+            foreach (Address daoAccount in DaoData.DaoAccounts)
+            {
+                UInt256 balance = _stateProvider.GetBalance(daoAccount);
+                _stateProvider.AddToBalance(withdrawAccount, balance, Dao.Instance);
+                _stateProvider.SubtractFromBalance(daoAccount, balance, Dao.Instance);
+            }
+        }
+    }
 }
